+++ src/main/java/oracle/fsgbu/eod/monitor/application/services/dto/IEodHistoryRepoModel.java
@@ -6,7 +6,7 @@ import java.util.Date;
 
 public interface IEodHistoryRepoModel {
 
-	Integer getMaxRunTime();
+	Double getMaxRunTime();
 
 	String getEocStage();
